+++ chaoslib/provider/process.py
@@ -8,13 +8,14 @@ from typing import Any
 from logzero import logger
 
 from chaoslib.exceptions import FailedActivity, InvalidActivity
-from chaoslib.types import Activity, Secrets
+from chaoslib.types import Activity, Configuration, Secrets
 
 
 __all__ = ["run_process_activity", "validate_process_activity"]
 
 
-def run_process_activity(activity: Activity, secrets: Secrets) -> Any:
+def run_process_activity(activity: Activity, configuration: Configuration,
+                         secrets: Secrets) -> Any:
     """
     Run the a process activity.