+++ chaoslib/rollback.py
@@ -5,14 +5,14 @@ from typing import Iterator
 from logzero import logger
 
 from chaoslib.activity import execute_activity
-from chaoslib.types import Experiment, Run, Secrets
+from chaoslib.types import Configuration, Experiment, Run, Secrets
 
 
 __all__ = ["run_rollbacks"]
 
 
-def run_rollbacks(experiment: Experiment, secrets: Secrets,
-                  pool: ThreadPoolExecutor,
+def run_rollbacks(experiment: Experiment, configuration: Configuration,
+                  secrets: Secrets, pool: ThreadPoolExecutor,
                   dry: bool = False) -> Iterator[Run]:
     """
     Run all rollbacks declared in the experiment in their order. Wait for
@@ -30,6 +30,8 @@ def run_rollbacks(experiment: Experiment, secrets: Secrets,
         if activity.get("background"):
             logger.debug("rollback activity will run in the background")
             yield pool.submit(execute_activity, activity=activity,
-                              secrets=secrets, dry=dry)
+                              configuration=configuration, secrets=secrets,
+                              dry=dry)
         else:
-            yield execute_activity(activity, secrets=secrets, dry=dry)
+            yield execute_activity(activity, configuration=configuration,
+                                   secrets=secrets, dry=dry)