+++ chaoslib/types.py
@@ -2,7 +2,8 @@
 from typing import Any, Dict, Generator, List, Tuple, Union
 
 __all__ = ["MicroservicesStatus", "Probe", "Action", "Experiment", "Layer",
-           "TargetLayers", "Activity", "Journal", "Run", "Secrets", "Step"]
+           "TargetLayers", "Activity", "Journal", "Run", "Secrets", "Step",
+           "Configuration"]
 
 
 Action = Dict[str, Any]
@@ -20,3 +21,4 @@ Run = Dict[str, Any]
 Step = Dict[str, Any]
 
 Secrets = Dict[str, Dict[str, str]]
+Configuration = Dict[str, Dict[str, str]]